requirements.txt
aiosqlite==0.21.0
aiohttp==3.11.14
async_lru==2.0.5
beautifulsoup4==4.13.4
dashscope==1.23.3
fake_useragent==2.2.0
fastapi==0.115.12
loguru==0.7.3
moviepy==2.2.1
openai==1.82.0
pandas==2.2.3
pycryptodome==3.23.0
pydantic==2.11.5
PyYAML==6.0.2
Requests==2.32.3
SQLAlchemy==2.0.41
streamlit==1.45.1
streamlit_authenticator==0.4.2
toml==0.10.2
uvicorn==0.34.2
